requirements.txt
fastapi
uvicorn
streamlit
requests
langchain
faiss-cpu
PyMuPDF
unstructured
python-docx
pyppt
ollama
tiktoken
langchain_community
python-multipart
langchain-community
langchain-ollama
pandas
pdfminer.six
sentence_transformers
torch
torchvision
torchaudio